requirements.txt
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
transformers>=4.12.0
diffusers>=0.16.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
ultralytics>=8.0.0
efficientnet-pytorch>=0.7.0
timm>=0.6.0
clip>=1.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
tqdm>=4.62.0
pyyaml>=5.4.0
einops>=0.4.0
accelerate>=0.20.0
huggingface-hub>=0.14.0